requirements.txt
djitellopy2==2.3
jupyter==1.0.0
numpy==1.19.5
pyserial
